Moved the TooManyPins modal into its forever home [RCORE-684]
