+++ app/src/molecules/Modal/OnDeviceDisplay/__tests__/TooManyPinsModal.test.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter } from 'react-router-dom'
 
 import { renderWithProviders } from '@opentrons/components'
 
-import { i18n } from '../../../i18n'
+import { i18n } from '../../../../i18n'
 import { TooManyPinsModal } from '../TooManyPinsModal'
 
 jest.mock('react-router-dom', () => {
