+++ app/src/molecules/Modal/OnDeviceDisplay/index.ts
@@ -1,0 +1,1 @@
+export { TooManyPinsModal } from './TooManyPinsModal'
